rename files and modify first run user check
