+++ lib/main.dart
@@ -13,16 +13,15 @@ import 'services/tts_manager.dart';
 import 'services/google_auth_service.dart';
 import 'services/firestore_manager.dart';
 import 'services/firebase_db.dart'; // walkydb 전용 db
-import 'services/UserDecider.dart'; // 판단 로직 가져오기
+import 'services/user_decider.dart'; // 판단 로직 가져오기
 
-import 'Onboarding_1.dart';
-import 'mainPage.dart';
+import 'onboarding_1.dart';
+import 'main_page.dart';
 
 void main() async {
   final widgetsBinding = WidgetsFlutterBinding.ensureInitialized();
   FlutterNativeSplash.preserve(widgetsBinding: widgetsBinding); // 앱이 준비될 때까지 스플래시 화면을 계속 유지하겠다고 선언
 
-  WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
@@ -31,13 +30,14 @@ void main() async {
   // : 기본적으로 Flutter에서 첫 프레임을 그리기 시작할 때 Splash Screen은 제거됨
   // 만약 앱이 초기화되는 동안에 Splash Screen을 유지하려면 preserve() 또는 remove() 메서드를 같이 사용하면 됨
 
-
   final bool isExistingUser; // 필드 선언
-  isExistingUser = await checkIsFirstRun();
+  isExistingUser = await isFirstRun();
 
   FlutterNativeSplash.remove(); // Splash 제거
   runApp(MyApp(isExistingUser: isExistingUser));
 }
+
+
 
 class MyApp extends StatelessWidget {
   final bool isExistingUser;
