+++ lib/onboarding_1.dart
@@ -7,7 +7,7 @@ class Onboarding_1 extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Main Page'),
+        title: const Text('Onboarding'),
       ),
       body: const Center(
           child: Text("이곳은 앱 다운로더 온보딩 페이지입니다.",
